requirements.txt
streamlit
tensorflow-cpu
numpy
pillow
keras
h5py
